Reorganize SlotsProxy unit tests
